+++ src/main/java/com/kdt/firststep/counselor/dto/request/CounselingReservationRequestDto.java
@@ -7,8 +7,8 @@ import java.time.LocalTime;
 
 @Getter
 public class CounselingReservationRequestDto {
-    private Long userId;              // 예약자 ID
-    private Long counselorId;         // 상담사 ID
+    private Integer userId;              // 예약자 ID
+    private Integer counselorId;         // 상담사 ID
     private LocalDate appointmentDate;  // 상담 날짜
     private LocalTime appointmentTime;  // 상담 시간
 }
